requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-appconf==1.2.0
django-apscheduler==0.6.1
django-cors-headers==4.9.0
django-embed-video==1.4.10
django-imagekit==6.0.0
django-unfold==0.74.1
djangorestframework==3.16.1
idna==3.11
pilkit==3.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
APScheduler==3.10.4
